+++ src/clojure/catacumba/handlers.clj
@@ -61,3 +61,14 @@
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; Basic request in Context
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn basic-request
+  "A chain handler that populates the context
+  with a basic request related properties."
+  [context]
+  (let [^Request request (:request context)]
+    (->> {:path (str "/" (.. request getPath))
+          :query-string (.. request getQuery)
+          :method (keyword (.. request getMethod getName toLowerCase))
+          :headers (ct/get-headers request)}
+         (ct/delegate context))))